#63 Make code refactoring of technologies for suitable overriding in enb-bem-i18n package
