+++ techs/bemhtml.js
@@ -23,78 +23,7 @@
  * ```
  */
 
-var path = require('path'),
-    vow = require('vow'),
-    vfs = require('enb/lib/fs/async-fs'),
-    pinpoint = require('pinpoint'),
-    XJST = require('bem-bl-xjst'),
-    BemhtmlProcessor = require('sibling').declare({
-        process: function (source, options) {
-            try {
-                return { result: XJST.translate(source, options) };
-            } catch (e) {
-                return { error: e };
-            }
-        }
-    }),
-    SourceMap;
-
-SourceMap = function (fileSources) {
-    var lastBoundary = 0,
-        boundaries = [],
-        filenames = [],
-        sources = [],
-        code = '';
-
-    fileSources.forEach(function (fileSource) {
-        var source = fileSource.source;
-
-        lastBoundary += source.split('\n').length;
-
-        boundaries.push(lastBoundary);
-        filenames.push(fileSource.filename);
-        sources.push(source);
-
-        code += '\n' + source;
-    });
-
-    function _getRangeIndexByLineNumber(lineNumber) {
-        if (lineNumber > lastBoundary + 1) {
-            return -1;
-        }
-
-        if (lineNumber < 0) {
-            return -1;
-        }
-
-        for (var i = 0; i < boundaries.length; ++i) {
-            if (lineNumber <= boundaries[i] + 1) {
-                return i;
-            }
-        }
-    }
-
-    return {
-        getOriginal: function (line, column) {
-            var rangeIndex = _getRangeIndexByLineNumber(line),
-                lowerBoundary = boundaries[rangeIndex - 1] || 0;
-
-            if (rangeIndex !== -1) {
-                return {
-                    filename: filenames[rangeIndex],
-                    source: sources[rangeIndex],
-                    line: line - lowerBoundary - 1,
-                    column: column
-                };
-            }
-        },
-        getCode: function () {
-            return code;
-        }
-    };
-};
-
-module.exports = require('enb/lib/build-flow').create()
+module.exports = require('./xjst').buildFlow()
     .name('bemhtml')
     .target('target', '?.bemhtml.js')
     .defineOption('exportName', 'BEMHTML')
@@ -102,76 +31,4 @@ module.exports = require('enb/lib/build-flow').create()
     .defineOption('devMode', true)
     .defineOption('cache', false)
     .useFileList('bemhtml')
-    .builder(function (sourceFiles) {
-        var _this = this,
-            node = this.node;
-        return vow.all(sourceFiles.map(function (file) {
-                return vfs.read(file.fullname, 'utf8')
-                    .then(function (source) {
-                        return {
-                            source: source,
-                            filename: file.fullname
-                        };
-                    });
-            }))
-            .then(function (fileSources) {
-                var bemhtmlProcessor = BemhtmlProcessor.fork(),
-                    sourceMap = SourceMap(fileSources),
-                    code = sourceMap.getCode();
-
-                node.getLogger().log('Calm down, OmetaJS is running...');
-
-                return bemhtmlProcessor.process(code, _this._getOptions())
-                    .then(function (res) {
-                        var result = res.result,
-                            error = res.error;
-
-                        bemhtmlProcessor.dispose();
-
-                        if (result) {
-                            return result;
-                        }
-
-                        if (error) {
-                            var line = error.line,
-                                column = error.column;
-
-                            // Syntax Error
-                            if (line && column) {
-                                var original = sourceMap.getOriginal(line, column);
-
-                                if (original) {
-                                    var message = error.message.split('\n')[0].replace(/\sat\:\s(\d)+\:(\d)+/, ''),
-                                        relPath = path.relative(node._root, original.filename),
-                                        context = pinpoint(original.source, {
-                                            line: original.line,
-                                            column: original.column,
-                                            indent: '    '
-                                        });
-
-                                    if (relPath.charAt(0) !== '.') {
-                                        relPath = './' + relPath;
-                                    }
-
-                                    throw new SyntaxError(message + ' at ' + relPath + '\n' + context);
-                                } else {
-                                    throw error;
-                                }
-                            } else {
-                                throw error;
-                            }
-                        }
-                    });
-            });
-    })
-    .methods({
-        _getOptions: function () {
-            return {
-                devMode: this._devMode,
-                cache: this._cache,
-                exportName: this._exportName,
-                applyFuncName: this._applyFuncName
-            };
-        }
-    })
     .createTech();
